// src/types.ts
export enum QUERIES {
  USERS = 'USERS',
  USER = 'USER',
  USER_POSTS = 'USER_POSTS',
  USER_ALBUMS = 'USER_ALBUMS',
  POST = 'POST',
  POSTS = 'POSTS',
  COMMENTS = 'COMMENTS',
}

export interface User {
  id: number;
  name: string;
  username: string;
  email: string;
  address: Address;
  phone: string;
  website: string;
  company: Company;
}

export interface Address {
  street: string;
  suite: string;
  city: string;
  zipcode: string;
  geo: Geo;
}

export interface Geo {
  lat: string;
  lng: string;
}

export interface Company {
  name: string;
  catchPhrase: string;
  bs: string;
}

export interface Post {
  userId: number;
  id: number;
  title: string;
  body: string;
}

export interface Comment {
  postId: number;
  id: number;
  name: string;
  email: string;
  body: string;
}